Rename Bend Image.png to Final Output.png
The output of the project is the final image (Output.jpg), where the rectangular pattern from Pattern.jpg is successfully mapped onto the curved surface of the flag image (Flag.jpg). Using OpenCV's image transformation techniques, such as perspective transformation and alpha blending, the pattern is warped to align with the natural folds and curves of the flag. The result is a realistic blend of the pattern that follows the flag’s movement, demonstrating the effectiveness of the applied image processing techniques.